fixed integer overflow bug and added flexible output.
Concretely, this commit adds a more versatile printTimings() function that
accepts a reference of a std::ostream. Hence, a user can either simply pass
std::cout, or they can open a file stream and pass it to the function if
they want the output to be dumped to a file.